Start splitting up booking process into smaller components
